% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char2NumVector.R
\name{char2NumVector}
\alias{char2NumVector}
\title{Character to Integer Vector}
\usage{
char2NumVector(c)
}
\arguments{
\item{c}{character vector.}
}
\value{
an integer vector.
}
\description{
Conversion from a character vector to an integer vector.
}
\examples{
c<-c("a","a","b")
num<-char2NumVector(c)
num

}
\author{
Ai-bing ZHANG, PhD. CNU, Beijing, CHINA. zhangab2008(at)mail.cnu.edu.cn
}
\references{
zhangab2008(at)mail.cnu.edu.cn
}
\keyword{char2NumVector}

