% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitize.DNA.R
\name{digitize.DNA}
\alias{digitize.DNA}
\title{Digitize DNAbin}
\usage{
digitize.DNA(seqs)
}
\arguments{
\item{seqs}{an object of DNAbin.}
}
\value{
a numeric matrix of DNA sequences digitized.
}
\description{
Digitize an object of DNAbin.
}
\examples{

data(TibetanMoth)
digitized.DNA<-digitize.DNA(seqs=TibetanMoth)
digitized.DNA


}
\author{
Ai-bing ZHANG, PhD. CNU, Beijing, CHINA.
}
\references{
zhangab2008(at)mail.cnu.edu.cn
}
\keyword{digitize.DNA}

