% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMES.R
\name{NAMES}
\alias{NAMES}
\title{Extracts Lables of Samples}
\usage{
NAMES(seqs)
}
\arguments{
\item{seqs}{object of class "DNAbin", generated from fasta2DNAbin() (package:adegenet), and read.dna() (package:ape).}
}
\value{
a character string array/vector.
}
\description{
Extract sequence names from different objects of DNAbin, including generated from fasta2DNAbin() (package:adegenet),
and read.dna() (package:ape).
}
\examples{

data(TibetanMoth) 
seqNames<-NAMES(TibetanMoth)
seqNames
}
\author{
Ai-bing ZHANG, PhD. CNU, Beijing, CHINA.
}
\references{
zhangab2008(at)mail.cnu.edu.cn
}
\keyword{NAMES}

