\name{pineMothCOI}
\alias{pineMothCOI}
\docType{data}
\title{
pine Moth COI
}
\description{
COI DNA barcodes of Pine Moth in China.
}
\usage{data("pineMothCOI")}
\format{
  The format is:
 'DNAbin' raw [1:140, 1:652] a t a a ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:140] "A111,Lasiocampidae_Dendrolimus_punctatus" "A22,Lasiocampidae_Dendrolimus_punctatus" "A23,Lasiocampidae_Dendrolimus_punctatus" "A27,Lasiocampidae_Dendrolimus_punctatus" ...
  ..$ : NULL
}
\details{
COI DNA barcodes of Pine Moth (Lasiocampidae) in China.
}
\source{
http://dx.plos.org/10.1371/journal.pone.0064428.
}
\references{
Dai Q-Y, Gao Q, Wu C-S, Chesters D, Zhu C-D, and A.B. Zhang*. (2012) Phylogenetic Reconstruction and DNA Barcoding for Closely Related Pine Moth Species (Dendrolimus) in China with Multiple Gene Markers. PLoS ONE 7(4): e32544.
}
\examples{
data(pineMothCOI)
pineMothCOI
}
\keyword{datasets}
