\name{pineMothITS1}
\alias{pineMothITS1}
\docType{data}
\title{
pine Moth ITS1
}
\description{
ITS1 sequences of seven closely related pine moths species sampled through China.
}
\usage{data("pineMothITS1")}
\format{
  The format is:
List of 69
 $ A43,Lasiocampidae_Dendrolimus_punctatus: raw [1:698] 48 48 48 28 ...
 $ A54,Lasiocampidae_Dendrolimus_punctatus: raw [1:696] 48 18 28 18 ...
 - attr(*, "class")= chr "DNAbin"
}
\details{
ITS1 used as DNA barcodes for closely related rine moth species (Dendrolimus) in China.
}
\source{
http://dx.plos.org/10.1371/journal.pone.0064428.
}
\references{
Dai Q-Y, Gao Q, Wu C-S, Chesters D, Zhu C-D, and A.B. Zhang*. (2012) Phylogenetic Reconstruction and DNA Barcoding for Closely Related Pine Moth Species (Dendrolimus) in China with Multiple Gene Markers. PLoS ONE 7(4): e32544.
}
\examples{
data(pineMothITS1)
pineMothITS1
}
\keyword{datasets}
