\name{pineMothITS2}
\alias{pineMothITS2}
\docType{data}
\title{
pine Moth ITS2
}
\description{
ITS2 sequences of seven closely related pine moths species sampled in China.
}
\usage{data("pineMothITS2")}
\format{
  The format is:
List of 97
 $ A22,Lasiocampidae_ Dendrolimus_ punctatus: raw [1:568] 48 48 48 28 ...
 $ A23,Lasiocampidae_ Dendrolimus_ punctatus: raw [1:574] 48 18 28 18 ...
 $ A29,Lasiocampidae_ Dendrolimus_ punctatus: raw [1:569] 88 48 48 48 ...
 $ A52,Lasiocampidae_ Dendrolimus_ punctatus: raw [1:570] 48 18 28 18 ...
 - attr(*, "class")= chr "DNAbin"
}
\details{
ITS2 used as DNA barcodes for closely related rine moth species (Dendrolimus) in China.
}
\source{
http://dx.plos.org /10.1371/ journal.pone.0064428.
}
\references{
Dai Q-Y, Gao Q, Wu C-S, Chesters D, Zhu C-D, and A.B. Zhang*. (2012) Phylogenetic Reconstruction and DNA Barcoding for Closely Related Pine Moth Species (Dendrolimus) in China with Multiple Gene Markers. PLoS ONE 7(4): e32544.
}
\examples{
data(pineMothITS2)
pineMothITS2
}
\keyword{datasets}
