\name{TibetanMoth}
\alias{TibetanMoth}
\docType{data}
\title{
Tibetan Moth
}
\description{
COI DNA barcodes of Tibetan Moth in China.
}
\usage{data("TibetanMoth")}
\format{
  The format is:
 'DNAbin' raw [1:319, 1:630] a t a a ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:319] "LS0909030M,Noctuidae_Himalaea_unica" "LZ0827026M,Noctuidae_Amphipyra_pyramidea" "ML0829010M,Noctuidae_Auchmis_saga" "BM0830055M,Noctuidae_Auchmis_saga" ...
  ..$ : NULL
}
\details{
COI DNA barcodes of Tibetan Moth (Noctuidae) in China.
}
\source{
http://dx.plos.org/10.1371/journal.pone.0064428.
}
\references{
Q. Jin, H.L. Han, X.M. Hu, X.H. Li,C.D. Zhu,S. Y. W. Ho, R. D. Ward, A.B. Zhang* . (2013).Quantifying Species Diversity with a DNA Barcoding-Based Method: Tibetan Moth Species (Noctuidae) on the Qinghai-Tibetan Plateau. PloS One 8: e644.
}
\examples{
data(TibetanMoth)
TibetanMoth
}
\keyword{datasets}
