% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.ref.R
\name{summarize.ref}
\alias{summarize.ref}
\title{Summarize Reference Data}
\usage{
summarize.ref(ref, taxonStat = TRUE, seqStat = TRUE, barcodeStat = TRUE)
}
\arguments{
\item{ref}{object of class "DNAbin" used as a reference dataset, which contains taxon information, or just an array
containing taxon information only.}

\item{taxonStat}{logic value to indicate whether the item is calculated only when ref is an object of class "DNAbin".}

\item{seqStat}{logic value to indicate whether the item is calculated}

\item{barcodeStat}{logic value to indicate whether the item is calculated}
}
\value{
a list containing taxon statistics, sequence statistics, population parameters,barcoding statistics ()
}
\description{
Summarize taxon information, sequence statistics,barcodes numbers per species for reference dataset.
}
\examples{

data(TibetanMoth)
s.r<-summarize.ref(TibetanMoth,taxonStat=TRUE,seqStat=TRUE,barcodeStat=TRUE)
s.r
}
\references{
zhangab2008(at)mail.cnu.edu.cn./zhangab2008(at)gmail.com.
}
\author{
Ai-bing ZHANG, Meng-di HAO, CNU, Beijing, CHINA.
}
\keyword{ref}
\keyword{summarize}
